Rearranged & Filtered the code
