+++ courses/static/js/add_course.js
@@ -82,7 +82,6 @@ function enableDropZone() {
     });
 }
 
-
 function enableReorder() {
     document.querySelectorAll(".builder-card").forEach(card => {
