+++ courses/templates/courses/instructor/add_event.html
@@ -64,7 +64,7 @@
             </div>
 
             <button type="submit" class="btn btn-success">💡 Create Event</button>
-            <a href="{% url 'instructor:instructor_dashboard' %}" class="btn btn-secondary">Cancel</a>
+            <a href="{% url 'instructor_dashboard' %}" class="btn btn-secondary">Cancel</a>
 
         </form>
