+++ courses/templates/courses/instructor/course_overview.html
@@ -12,7 +12,7 @@
 
     <!-- HEADER -->
     <div class="d-flex justify-content-between align-items-center p-3 bg-white shadow-sm">
-        <a href="{% url 'instructor:instructor_dashboard' %}" class="text-decoration-none text-dark">
+        <a href="{% url 'instructor_dashboard' %}" class="text-decoration-none text-dark">
             <h3 class="m-0 fw-bold">Bildung</h3>
         </a>
 
@@ -28,7 +28,7 @@
                 <li><a class="dropdown-item" href="{% url 'instructor:add_course' %}">Create New Course</a></li>
                 <li><a class="dropdown-item" href="{% url 'instructor:students_list' %}">My Students</a></li>
                 <li><a class="dropdown-item" href="{% url 'instructor:calendar_view' %}">My Schedules</a></li>
-                <li><a class="dropdown-item" href="{% url 'instructor:instructor_notifications' %}">Notifications</a></li>
+                <li><a class="dropdown-item" href="{% url 'instructor_notifications' %}">Notifications</a></li>
                 <li><a class="dropdown-item" href="#">Earnings & Payments</a></li>
                 <li><a class="dropdown-item" href="#">Help & Support</a></li>
                 <li><a class="dropdown-item" href="{% url 'instructor:account_settings' %}">Account Settings</a></li>
