--- courses/templates/courses/instructor/create_course.html
@@ -1,10 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-<h2>Create New Course</h2>
-<form method="post">
-    {% csrf_token %}
-    {{ form.as_p }}
-    <button type="submit">Save Course</button>
-</form>
-<a href="{% url 'instructor_dashboard' %}">Back to Dashboard</a>
-{% endblock %}
