+++ courses/templates/courses/instructor/edit_course.html
@@ -130,7 +130,7 @@
 
 <div class="topbar">
 
-    <a href="{% url 'instructor:instructor_dashboard' %}" class="topbar-title">
+    <a href="{% url 'instructor_dashboard' %}" class="topbar-title">
         Bildung
     </a>
 
@@ -147,7 +147,7 @@
                 <li><a class="dropdown-item" href="{% url 'instructor:add_course' %}">Create New Course</a></li>
                 <li><a class="dropdown-item" href="{% url 'instructor:students_list' %}">My Students</a></li>
                 <li><a class="dropdown-item" href="{% url 'instructor:calendar_view' %}">My Schedules</a></li>
-                <li><a class="dropdown-item" href="{% url 'instructor:instructor_notifications' %}">Notifications</a></li>
+                <li><a class="dropdown-item" href="{% url 'instructor_notifications' %}">Notifications</a></li>
                 <li><a class="dropdown-item" href="#">Earnings & Payments</a></li>
                 <li><a class="dropdown-item" href="#">Help & Support</a></li>
                 <li><a class="dropdown-item" href="{% url 'instructor:account_settings' %}">Account Settings</a></li>
