--- courses/templates/courses/instructor/my_activity.html
@@ -1,82 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<meta charset="UTF-8">
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
-<title>My activity</title>
-<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
-
-</head>
-<body class="bg-light">
-  <div class="d-flex justify-content-between align-items-center p-3 bg-white shadow-sm">
-    <a href="{% url 'instructor:instructor_dashboard' %}" class="text-decoration-none text-dark">
-        <h3 class="m-0 fw-bold">Bildung</h3>
-    </a>  
-
-        <div class="dropdown">
-            <button class="btn rounded-circle shadow" style="background:gray;color:white;width:45px;height:45px;" type="button" data-bs-toggle="dropdown">
-                {{ request.user.username|first|upper }}
-            </button>
-            <ul class="dropdown-menu dropdown-menu-end">
-                <li><a class="dropdown-item" href="{% url 'instructor_profile_view' %}">Profile</a></li>
-                <li><a class="dropdown-item" href="{% url 'instructor:add_course' %}">Create New Course</a></li>
-                <li><a class="dropdown-item" href="{% url 'instructor:my_activity' %}">My Activity</a></li>
-                <li><a class="dropdown-item" href="{% url 'instructor:students_list' %}">My Students</a></li>
-                <li><a class="dropdown-item" href="{% url 'instructor:calendar_view' %}">My Schedules</a></li>
-                <li><a class="dropdown-item" href="{% url 'instructor:instructor_notifications' %}">Notifications</a></li>
-                <li><a class="dropdown-item" href="#">Earnings & Payments</a></li>
-                <li><a class="dropdown-item" href="#">Help & Support</a></li>
-                <li><a class="dropdown-item" href="{% url 'instructor:account_settings' %}">Account Settings</a></li>
-                <li><hr class="dropdown-divider"></li>
-                <li><a class="dropdown-item text-danger" href="{% url 'logout_view' %}">Logout</a></li>
-            </ul>
-        </div>
-    </div>
-</div>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-<div class="container py-5">
-  <h2 class="fw-bold text-primary mb-4">🎥 My Scheduled Live Classes</h2>
-  
-  {% if live_classes %}
-  <table class="table table-striped shadow-sm rounded-3">
-    <thead class="table-primary">
-      <tr>
-        <th>Course</th>
-        <th>Topic</th>
-        <th>Date</th>
-        <th>Time</th>
-        <th>Join</th>
-      </tr>
-    </thead>
-
-    <tbody>
-      {% for session in live_classes %}
-        <tr>
-          <td>{{ session.course.title }}</td>
-          <td>{{ session.topic }}</td>
-          <td>{{ session.date }}</td>
-          <td>{{ session.time }}</td>
-          <td>
-            {% if session.meeting_link %}
-               <a href="{{ session.meeting_link }}" target="_blank" class="btn btn-success btn-sm">
-              Join </a>
-            {% else %}
-                 <a href="" onclick="return false;" class="btn btn-secondary btn-sm">
-                No Link
-                    </a>
-            {% endif %}
-
-          </td>
-
-
-        </tr>
-      {% endfor %}
-    </tbody>
-  </table>
-{% else %}
-  <div class="alert alert-info">No live classes scheduled yet.</div>
-{% endif %}
-
-</div>
-</body>
-</html>
