+++ courses/templates/courses/instructor/schedule_live_class.html
@@ -10,7 +10,7 @@
 <body class="bg-light">
 
   <div class="d-flex justify-content-between align-items-center p-3 bg-white shadow-sm">
-    <a href="{% url 'instructor:instructor_dashboard' %}" class="text-decoration-none text-dark">
+    <a href="{% url 'instructor_dashboard' %}" class="text-decoration-none text-dark">
         <h3 class="m-0 fw-bold">Bildung</h3>
     </a>
 
@@ -26,7 +26,7 @@
                 <li><a class="dropdown-item" href="{% url 'instructor:add_course' %}">Create New Course</a></li>
                 <li><a class="dropdown-item" href="{% url 'instructor:students_list' %}">My Students</a></li>
                 <li><a class="dropdown-item" href="{% url 'instructor:calendar_view' %}">My Schedules</a></li>
-                <li><a class="dropdown-item" href="{% url 'instructor:instructor_notifications' %}">Notifications</a></li>
+                <li><a class="dropdown-item" href="{% url 'instructor_notifications' %}">Notifications</a></li>
                 <li><a class="dropdown-item" href="#">Earnings & Payments</a></li>
                 <li><a class="dropdown-item" href="#">Help & Support</a></li>
                 <li><a class="dropdown-item" href="{% url 'instructor:account_settings' %}">Account Settings</a></li>
@@ -47,7 +47,7 @@
 
 
   <button type="submit" class="btn btn-success">💡 Schedule</button>
-  <a href="{% url 'instructor:instructor_dashboard' %}" class="btn btn-secondary">Cancel</a>
+  <a href="{% url 'instructor_dashboard' %}" class="btn btn-secondary">Cancel</a>
 </form>
 
       </div>
