--- courses/templates/courses/instructor_course_detail.html
@@ -1,31 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-<h2>{{ course.title }}</h2>
-<p>{{ course.description }}</p>
-
-<h3>Your Lectures</h3>
-<ul>
-    {% for lecture in course.lectures.all %}
-        <li>{{ lecture.title }}</li>
-    {% empty %}
-        <li>No lectures yet.</li>
-    {% endfor %}
-</ul>
-
-<h3>Feedbacks Given</h3>
-<ul>
-    {% for feedback in course.feedbacks.all %}
-        <li>
-            To: {{ feedback.student.username }} <br>
-            {{ feedback.text }} <br>
-            <small>{{ feedback.created_at }}</small>
-        </li>
-    {% empty %}
-        <li>No feedback given yet.</li>
-    {% endfor %}
-</ul>
-
-
-<a href="{% url 'add_lecture' course.id %}">+ Add Lecture</a> |
-<a href="{% url 'give_feedback' course.id %}">Give Feedback</a>
-{% endblock %}
