--- courses/templates/courses/student/course_detail.html
@@ -1,36 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-<div class="container mt-4">
-  <h3>{{ course.title }}</h3>
-  <p>{{ course.description }}</p>
-  <hr>
-
-  <h5>Modules & Lectures</h5>
-  {% for module in modules %}
-    <div class="card mb-3">
-      <div class="card-header bg-light">
-        <strong>{{ module.title }}</strong>
-      </div>
-      <div class="card-body">
-        {% for lecture in module.lectures.all %}
-          <div class="d-flex justify-content-between align-items-center border-bottom py-2">
-            <span>{{ lecture.title }}</span>
-            {% if progress_map|get_item:lecture.id %}
-              <span class="badge bg-success">Completed</span>
-            {% else %}
-              <a href="{% url 'student:mark_lecture_complete' lecture.id %}" class="btn btn-outline-primary btn-sm">Mark Complete</a>
-            {% endif %}
-          </div>
-        {% empty %}
-          <p>No lectures in this module.</p>
-        {% endfor %}
-      </div>
-    </div>
-  {% endfor %}
-
-  <div class="mt-3">
-    <p><strong>Completed:</strong> {{ completed }} / {{ total }} lectures</p>
-    <a href="{% url 'student:student_course_progress' course.id %}" class="btn btn-secondary">View Progress Report</a>
-  </div>
-</div>
-{% endblock %}
