+++ users/templates/instructor/instructor_login.html
@@ -1,0 +1,74 @@
+{% extends "base.html" %}
+{% block content %}
+<div class="container mt-5">
+  <div class="col-md-6 offset-md-3">
+    <div class="card shadow p-4">
+      <h2>Instructor Login</h2>
+      
+    {% if messages %}
+  <div aria-live="polite" aria-atomic="true" class="position-relative">
+    <div class="toast-container position-fixed top-0 end-0 p-3" style="z-index: 9999;">
+      {% for message in messages %}
+        <div class="toast align-items-center text-white bg-{{ message.tags }} border-0"
+             role="alert" aria-live="assertive" aria-atomic="true" data-bs-delay="5000"
+             id="autoToast">
+          <div class="d-flex">
+            <div class="toast-body">
+              {{ message }}
+            </div>
+            <button type="button" class="btn-close btn-close-white me-2 m-auto"
+                    data-bs-dismiss="toast"></button>
+          </div>
+        </div>
+      {% endfor %}
+    </div>
+  </div>
+  <script>
+    document.addEventListener("DOMContentLoaded", function () {
+        var toastEl = document.getElementById("autoToast");
+        if (toastEl) {
+            var toast = new bootstrap.Toast(toastEl);
+            toast.show();
+
+            setTimeout(function() {
+                const msg = toastEl.innerText.toLowerCase();
+
+                if (msg.includes("login")) {
+                    window.location.href = "{% url 'student_dashboard' %}";
+                }
+
+                if (msg.includes("logout")) {
+                    window.location.href = "{% url 'auth_page' %}";
+                }
+            }, 5000);
+        }
+    });
+  </script>
+
+{% endif %}
+
+      <form method="post">
+        {% csrf_token %}
+        {{ form.as_p }}
+        <button type="submit" class="btn btn-success w-100">Login</button>
+      </form>
+
+      <p class="mt-2 text-center">
+        <a href="{% url 'forgot_password' %}">Forgot Password?</a>
+      </p>
+
+      <div class="text-center my-3">
+        <span class="text-muted">Or continue with</span>
+        <hr class="mt-2 mb-3">
+        <a href="{% url 'google_oauth_entry' %}?type=instructor" class="btn btn-outline-danger w-100">
+          <i class="fab fa-google"></i> Continue with Google
+        </a>
+      </div>
+
+      <p class="mt-3">Don't have an account? 
+        <a href="{% url 'instructor_signup' %}">Instructor Signup</a>
+      </p>
+    </div>
+  </div>
+</div>
+{% endblock %}
