+++ users/templates/instructor/instructor_notifications.html
@@ -27,7 +27,7 @@
 <!-- HEADER -->
 <div class="d-flex justify-content-between align-items-center p-3 bg-white shadow-sm">
 
-    <a href="{% url 'instructor:instructor_dashboard' %}" class="text-decoration-none text-dark">
+    <a href="{% url 'instructor_dashboard' %}" class="text-decoration-none text-dark">
         <h3 class="m-0 fw-bold">Bildung</h3>
     </a>
 
@@ -40,13 +40,15 @@
 
         <ul class="dropdown-menu dropdown-menu-end">
             <li><a class="dropdown-item" href="{% url 'instructor_profile_view' %}">Profile</a></li>
-            <li><a class="dropdown-item" href="{% url 'instructor:add_course' %}">Create New Course</a></li>
-            <li><a class="dropdown-item" href="{% url 'instructor:students_list' %}">My Students</a></li>
-            <li><a class="dropdown-item" href="{% url 'instructor:calendar_view' %}">My Schedules</a></li>
-            <li><a class="dropdown-item" href="{% url 'instructor:instructor_notifications' %}">Notifications</a></li>
-            <li><a class="dropdown-item" href="{% url 'instructor:account_settings' %}">Account Settings</a></li>
-            <li><hr class="dropdown-divider"></li>
-            <li><a class="dropdown-item text-danger" href="{% url 'logout_view' %}">Logout</a></li>
+                <li><a class="dropdown-item" href="{% url 'instructor:add_course' %}">Create New Course</a></li>
+                <li><a class="dropdown-item" href="{% url 'instructor:students_list' %}">My Students</a></li>
+                <li><a class="dropdown-item" href="{% url 'instructor:calendar_view' %}">My Schedules</a></li>
+                <li><a class="dropdown-item" href="{% url 'instructor_notifications' %}">Notifications</a></li>
+                <li><a class="dropdown-item" href="#">Earnings & Payments</a></li>
+                <li><a class="dropdown-item" href="#">Help & Support</a></li>
+                <li><a class="dropdown-item" href="{% url 'instructor:account_settings' %}">Account Settings</a></li>
+                <li><hr class="dropdown-divider"></li>
+                <li><a class="dropdown-item text-danger" href="{% url 'logout_view' %}">Logout</a></li>
         </ul>
     </div>
 </div>
@@ -103,12 +105,12 @@
 <!-- JS -->
 <script>
 function markRead(id) {
-    fetch("{% url 'instructor:mark_notification_read' 0 %}".replace("0", id))
+    fetch("{% url 'mark_notification_read' 0 %}".replace("0", id))
         .then(() => location.reload());
 }
 
 function markAllRead() {
-    fetch("{% url 'instructor:mark_notifications_read' %}")
+    fetch("{% url 'mark_notifications_read' %}")
         .then(() => location.reload());
 }
 </script>
