--- users/templates/users/student_signup.html
@@ -1,104 +1,0 @@
-{% load static %}
-
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-
-  <!-- Your CSS -->
-  <link rel="stylesheet" type="text/css" href="{% static 'inst_signup/style.css' %}">
-
-  <!-- Font Awesome for icons -->
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
-
-  <title>Student Registration</title>
-</head>
-<body>
-  <nav class="container">
-    <a href="/">Home</a>
-  </nav>
-
-  <div class="container">
-    <div class="header">
-      <h1>Create Your Account</h1>
-      <p>Join our learning community as a Student</p>
-    </div>
-
-    <div class="container mt-5">
-      <div class="col-md-6 offset-md-3">
-        <div class="card shadow p-4">
-
-          <form method="post">
-            {% csrf_token %}
-
-            <div class="row">
-              <div class="form-group">
-                {{ form.first_name.label_tag }}
-                {{ form.first_name }}
-                {% if form.first_name.errors %}{{ form.first_name.errors }}{% endif %}
-              </div>
-              <div class="form-group">
-                {{ form.last_name.label_tag }}
-                {{ form.last_name }}
-                {% if form.last_name.errors %}{{ form.last_name.errors }}{% endif %}
-              </div>
-            </div>
-
-            <div class="form-group">
-              {{ form.email.label_tag }}
-              {{ form.email }}
-              {% if form.email.errors %}{{ form.email.errors }}{% endif %}
-            </div>
-
-            <div class="form-group">
-              {{ form.password1.label_tag }}
-              {{ form.password1 }}
-              <i class="fa-solid fa-eye toggle-password" data-target="{{ form.password1.id_for_label }}"></i>
-              {% if form.password1.errors %}{{ form.password1.errors }}{% endif %}
-            </div>
-
-            <div class="form-group">
-              {{ form.password2.label_tag }}
-              {{ form.password2 }}
-              <i class="fa-solid fa-eye toggle-password" data-target="{{ form.password2.id_for_label }}"></i>
-              {% if form.password2.errors %}{{ form.password2.errors }}{% endif %}
-            </div>
-
-            <button type="submit" class="btn btn-success w-100">Sign Up</button>
-          </form>
-
-          <!-- Added Google Signup Button (Only Change) -->
-          <div class="text-center my-3">
-            <span class="text-muted">Or sign up with</span>
-            <hr class="mt-2 mb-3">
-            <a href="{% url 'google_oauth_entry' %}?type=student" class="btn btn-outline-danger w-100">
-              <i class="fab fa-google"></i> Continue with Google
-            </a>
-          </div>
-          <!-- End Google Signup Section -->
-
-          <p class="mt-3">Already have an account? 
-            <a href="{% url 'student_login' %}">Student Login</a>
-          </p>
-        </div>
-      </div>
-    </div>
-  </div>
-
-  <script>
-    // Toggle password visibility
-    document.querySelectorAll(".toggle-password").forEach((icon) => {
-      icon.addEventListener("click", function () {
-        const targetId = this.getAttribute("data-target");
-        const input = document.getElementById(targetId);
-        const type = input.getAttribute("type") === "password" ? "text" : "password";
-        input.setAttribute("type", type);
-        this.classList.toggle("fa-eye");
-        this.classList.toggle("fa-eye-slash");
-      });
-    });
-  </script>
-
-</body>
-</html>
